Added Time spec RSC16
