+++ src/IO.Ably.Tests/Rest/RestSandBoxSpecs.cs
@@ -1,0 +1,50 @@
+using System;
+using System.Threading.Tasks;
+using FluentAssertions;
+using Xunit;
+
+namespace IO.Ably.Tests
+{
+    [Collection("AblyRest SandBox Collection")]
+    public class RestSandBoxSpecs
+    {
+        private readonly AblySandboxFixture _fixture;
+
+        public RestSandBoxSpecs(AblySandboxFixture fixture)
+        {
+            _fixture = fixture;
+        }
+
+        private async Task<AblyRest> GetRestClient(Protocol protocol)
+        {
+            var settings = await _fixture.GetSettings();
+            var defaultOptions = settings.CreateDefaultOptions();
+            defaultOptions.UseBinaryProtocol = protocol == Protocol.MsgPack;
+            return new AblyRest(defaultOptions);
+        }
+
+        [Theory]
+        [ProtocolData]
+        [Trait("spec", "RSC6a")]
+        public async Task GettingStats_ShouldReturnValidPaginatedResultOfStats(Protocol protocol)
+        {
+            var client = await GetRestClient(protocol);
+
+            var stats = await client.Stats(new StatsDataRequestQuery());
+
+            stats.Should().NotBeNull();
+        }
+
+        [Theory]
+        [ProtocolData]
+        [Trait("spec", "RSC16")]
+        public async Task Time_ShouldReturnAValidDateTimeOffset(Protocol protocol)
+        {
+            var client = await GetRestClient(protocol);
+
+            var now = await client.Time();
+
+            now.Should().BeCloseTo(Config.Now(), (int)TimeSpan.FromHours(1).TotalMilliseconds);
+        }
+    }
+}
